fix a bug in the run_single_covid.sh script
the wrong argument was being passed for the number of simulations to run.
this is now fixed, the script is reworded to make this more explicit
(fewer magic numbers), and some documentation on the command-line arguments
has been added to UK.R to prevent a similar error from occurring in future.